requirements.txt
python-telegram-bot[webhooks]
google-generativeai
tweepy
python-dotenv
urllib3<2
importlib-metadata
APScheduler
